Implemented exception handlers for method deleteById and created a DatabaseException class to handle database integrity exceptions

+++ src/main/java/com/pursuit/springclass/services/UserService.java
@@ -3,10 +3,15 @@ package com.pursuit.springclass.services;
 import com.pursuit.springclass.entities.User;
 import com.pursuit.springclass.repositories.UserRepository;
 import com.pursuit.springclass.resources.exceptions.ResourceExceptionHandler;
+import com.pursuit.springclass.services.exceptions.DatabaseException;
 import com.pursuit.springclass.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.PersistenceException;
 import jakarta.servlet.http.HttpServletRequest;
+import org.h2.jdbc.JdbcSQLIntegrityConstraintViolationException;
+import org.hibernate.exception.ConstraintViolationException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpServerErrorException;
 
 import java.util.List;
 import java.util.Optional;
@@ -33,7 +38,15 @@ public class UserService {
     }
 
     public void deleteById(Long id) {
-        this.userRepository.deleteById(id);
+        if(this.userRepository.findById(id).isEmpty()) {
+            throw new ResourceNotFoundException(id);
+        }
+
+        try {
+            this.userRepository.deleteById(id);
+        } catch(RuntimeException e) {
+            throw new DatabaseException("It is not possible to delete that user, because there is another entity containing a foreign key that points to the specified user.");
+        }
     }
 
     public User update(Long id, User obj) {
